Update to allow for android-sdk-install (via jenkins job)
Motivation:
The current android slave docker container has the android sdk 'baked' into the image
To avoid any legal problems with the license agreement it would be better to exclude the install and setup of the sdk in a container

Solution:
The best solution (and easiest - after days of trying plugins/master-slave configurations with external directory mounts etc) was found to be :-

  - Create a simple freestyle jenkins job
  - Copy the script (included in this repo jenkinsfiles/install-android-sdk.sh) into the job execute shell field
  - Add a Global variable GRADLE_HOME that points to the gradle wrapper directory
  - Create the android slave build pipeline with the script example (jenkinsfiles/new-android.groovy)
  - The install android job needs to be executed once only

Simple changes to the current dockerfile were necessary (set the JAVA_HOME env for gradle and include wget for the install script)

Result:
The slave relies on the GRADLE_HOME setting and calls the gradlew script.
This has been tested locally.

JIRA:
refer to AGDIGGER-43 for details and logs