Update to create an s2i build for the android sdk
Motivation:
With the jenkins openshift platform the mounting of an external PV for the android sdk for each instance of a slave proved to be problamatic.
A new approach is to use a s2i image to include the installed sdk.
The thinking here was to allow the admin user to download install and accept the license agreement.
The s2i build would then include the sdk into a s2i slave base image. The resultant image would then be uploaded to the internal docker openshift registry.

Changes:
A new directory structure was included into the repo , with all the valid scripts and Dockerfile to realize this functionality

Goals:
The final goal is to use the slave base s2i image within an ansible script to automate various versions of the android sdk

JIRA:
Refer to https://issues.jboss.org/browse/AGDIGGER-63